requirements.txt
pydantic
python-dotenv
# django and it's extensions
Django
django-debug-toolbar
django-extensions
mysqlclient==2.1.1
# aiogram module from dev-3.x branch
https://github.com/aiogram/aiogram/archive/refs/heads/dev-3.x.zip
djangorestframework~=3.14.0
SQLAlchemy~=2.0.15
requests~=2.31.0